Optimized the SurfaceTagWeights data structure. The optimization required imposing a fixed limit of 16 weights per point, which seems like a reasonable limit (most points will likely have 1-2 weights). At 16 points, the structure takes 132 bytes, which is still smaller than the 176 bytes previously used by an empty unordered_map.

While changing it, I also switched it from using a pair<> to using the AzFramework::SurfaceTagWeight as the underlying data type. I also noticed that there was undefined behavior when adding the same tag twice, so I defined it to always preserve the highest weight, which made the "AddSurfaceWeightIfGreater()" redundant with "AddSurfaceWeight()", so I merged the two APIs together.

This also adds benchmarks for the AddSurfaceWeight and HasMatchingTags APIs, which are the two most performance-critical, so that it's possible to verify that changing from an unordered_map to a fixed_vector is an improvement.

Before (unordered_map):
```
------------------------------------------------------------------------------------------------------
Benchmark                                                            Time             CPU   Iterations
------------------------------------------------------------------------------------------------------
SurfaceDataBenchmark/BM_GetSurfacePoints/1024                     1201 ms         1203 ms            1
SurfaceDataBenchmark/BM_GetSurfacePoints/2048                     4771 ms         4766 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/1024           1750 ms         1750 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/2048           7086 ms         7094 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/1024             1754 ms         1750 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/2048             6556 ms         6563 ms            1
SurfaceDataBenchmark/BM_AddSurfaceTagWeight/ClearEachTime:0        215 ns          215 ns      3200000
SurfaceDataBenchmark/BM_AddSurfaceTagWeight/ClearEachTime:1        963 ns          963 ns       746667
SurfaceDataBenchmark/BM_HasAnyMatchingTags_NoMatches               107 ns          105 ns      6400000
```

After (fixed_vector):
```
------------------------------------------------------------------------------------------------------
Benchmark                                                            Time             CPU   Iterations
------------------------------------------------------------------------------------------------------
SurfaceDataBenchmark/BM_GetSurfacePoints/1024                      680 ms          688 ms            1
SurfaceDataBenchmark/BM_GetSurfacePoints/2048                     2737 ms         2734 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/1024           1167 ms         1172 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/2048           4696 ms         4688 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/1024             1176 ms         1188 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/2048             4462 ms         4453 ms            1
SurfaceDataBenchmark/BM_AddSurfaceTagWeight/ClearEachTime:0        127 ns          128 ns      5600000
SurfaceDataBenchmark/BM_AddSurfaceTagWeight/ClearEachTime:1        134 ns          132 ns      4977778
SurfaceDataBenchmark/BM_HasAnyMatchingTags_NoMatches              60.2 ns         60.9 ns     10000000
```